requirements.txt

openai>=1.0.0
tiktoken>=0.5.2
faiss-cpu>=1.8.0
streamlit>=1.36.0
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.1
pydantic>=2.6.0
